change: 一般的な買い物では消費税の端数は切り捨てられる場合が多いため、価格計算では小数点以下は切り捨てるよう変更

+++ app/src/main/java/net/zuuno/regicalc/presentation/calculator/CalculatorUiState.kt
@@ -8,7 +8,7 @@ data class CalculatorUiState(
     val price: String = "",
     val quantity: String = "",
     val operation: CalculatorOperation? = null,
-    val totalPrice: Double = 0.0,
+    val totalPrice: Int = 0,
     val shoppingList: List<Shopping> = emptyList()
 )
