+++ app/src/main/java/net/zuuno/regicalc/presentation/calculator/util/Shopping.kt
@@ -4,12 +4,12 @@ import java.time.LocalDateTime
 
 data class Shopping(
     val id: String,
-    val price: Double,
+    val price: Int,
     val quantity: Int,
     val createdAt: LocalDateTime = LocalDateTime.now(),
     val taxRate: TaxRate = TaxRate.Default
 )
 
-fun Shopping.totalPrice(): Double {
-    return price * quantity * (1 + taxRate.rate)
+fun Shopping.totalPrice(): Int {
+    return( price * quantity * (1 + taxRate.rate)).toInt()
 }
